homework 8 tips
You need the widgets parameter in the function. (The widgets will be updated by function)
Function add inventory, If widgets not in inventory then else assumes it is, no need for if check again.
Function add inventory, In the else statement you must update the value too.

IN remove inventory you need the widgets parameter, again the widgets variable for dictionary will be updated in function.